// Content/app/generated/ModuleMarkdown.ts
// This file is auto generated, don't edit it manually.
export const ModuleMarkdown = {
  Dashboard: {
    components: {
      'main-nav': {
        'label-dashboard': {
          content: 'Servers',
          documentation: '',
        },
        'label-home': {
          content: 'Home',
          documentation: '',
        },
        'label-logout': {
          content: 'Logout',
          documentation: '',
        },
        'label-modules': {
          content: 'Modules',
          documentation: '',
        },
        'label-overview': {
          content: 'Overview',
          documentation: '',
        },
        'label-profile': {
          content: 'Profile',
          documentation: '',
        },
      },
    },
    pages: {
      dashboard: {
        title: {
          content: 'Servers',
          documentation: '',
        },
      },
      home: {
        'add-button': {
          content: 'Add To Server',
          documentation: '',
        },
        'join-button': {
          content: 'Join us on Discord',
          documentation: '',
        },
        'landing-description': {
          content:
            'Based on your community & everything that involves it. I try to keep my communities active, safe, and fun for everyone, and this without any product placement.',
          documentation: '',
        },
        'landing-title': {
          content: 'Your community, our duty!',
          documentation: '',
        },
        title: {
          content: 'Home',
          documentation: '',
        },
      },
      profile: {
        title: {
          content: 'Profile',
          documentation: '',
        },
      },
    },
  },
  Global: {
    months: {
      content:
        'January\nFebruary\nMarch\nApril\nMay\nJune\nJuly\nAugust\nSeptember\nOctober\nNovember\nDecember',
      documentation: '',
    },
  },
};

export default ModuleMarkdown;